balanced field ODEs now share common functions
